+++ packages/lucia-auth/src/auth/user/mutate.ts
@@ -1,6 +1,32 @@
 import { hashScrypt } from "../../utils/crypto.js";
 import type { Auth, User } from "../../types.js";
 
+type CreateUser = (
+	provider: string,
+	identifier: string,
+	options?: {
+		password?: string;
+		attributes?: Lucia.UserAttributes;
+	}
+) => Promise<User>;
+
+export const createUserFunction = (auth: Auth) => {
+	const createUser: CreateUser = async (provider, identifier, options) => {
+		const providerId = `${provider}:${identifier}`;
+		const attributes = options?.attributes || {};
+		const userId = await auth.configs.generateCustomUserId();
+		const hashedPassword = options?.password ? await hashScrypt(options.password) : null;
+		const userData = await auth.configs.adapter.setUser(userId, {
+			providerId,
+			hashedPassword: hashedPassword,
+			attributes
+		});
+		const user = auth.configs.transformUserData(userData);
+		return user;
+	};
+	return createUser;
+};
+
 type UpdateUserProviderIdToken = (
 	userId: string,
 	provider: string,
@@ -50,4 +76,14 @@ export const updateUserPasswordFunction = (auth: Auth) => {
 		return user;
 	};
 	return updateUserPassword;
+};
+
+type DeleteUser = (userId: string) => Promise<void>;
+
+export const deleteUserFunction = (auth: Auth) => {
+	const deleteUser: DeleteUser = async (userId: string) => {
+		await auth.configs.adapter.deleteSessionsByUserId(userId);
+		await auth.configs.adapter.deleteUser(userId);
+	};
+	return deleteUser;
 };